00-Starter-Seed/requirements.txt
flask
python-dotenv
requests
flask-oauthlib